Move steppers out of the engine into a game specific folder
