+++ lib/panel-league/basic.js
@@ -1,5 +1,15 @@
-const JKISS31 = require('./jkiss');
-const {newBlock, clearBlock, floodFill, shuffleInPlace} = require('./engine-util');
+const JKISS31 = require('../jkiss');
+const {newBlock, clearBlock, floodFill, shuffleInPlace} = require('./util');
+
+// Block properties
+// Non-solid: Null color blocks act as air.
+// Solid: The block can be swapped with other blocks and air and is displayed displayed by the UI.
+// Floating: If a block has air under it, it will start to float and hang in the air for a while (floatTimer > 1).
+// Falling: The float timer has ran out and the block is in free fall one unit per time step (floatTimer in {0, 1}).
+// Landed: The block has a solid block beneath it and is no longer in free fall (floatTimer < 0).
+// Flashing: The block is part of a match (of 3 or more) and will soon turn into air (flashTimer >= 0).
+// Chaining: The block is part of a continuous chain where blocks from a previous match drop into a new match (chaining == true).
+// Swapping: The block has begun swapping and is on its way to the new location (swapTimer != 0).
 
 const blocksCanSwap = ((block1, block2, above1, above2) => {
   if (!block1 || !block2) {
