+++ lib/panel-league/constant.js
