+++ lib/panel-league/garbage.js
@@ -1,5 +1,5 @@
-const JKISS31 = require('./jkiss');
-const {floodFill} = require('./engine-util.js');
+const JKISS31 = require('../jkiss');
+const {floodFill} = require('./util');
 
 const garbageCoordsToIndex = ((state, x, y) => {
   if (x < 0) {
