+++ lib/panel-league/util.js
